+++ class/subtotal.class.php
@@ -101,10 +101,11 @@ class TSubtotal {
 	 * @param string       $label
 	 * @param int          $qty
 	 * @param int          $rang
+	 * @param string $desc
 	 * @return int
 	 *
 	 */
-	static function addSubTotalLine(&$object, $label, $qty, $rang=-1) {
+	static function addSubTotalLine(&$object, $label, $qty, $rang=-1, $desc = '') {
 
 		$res = 0;
 
@@ -124,7 +125,6 @@ class TSubtotal {
 
 		}
 		else {
-			$desc = '';
 
 			$TNotElements = array ('invoice_supplier', 'order_supplier');
 			if ((float) DOL_VERSION < 6  || $qty==50 && !in_array($object->element, $TNotElements)) {
@@ -269,16 +269,17 @@ class TSubtotal {
 		}
 	}
 
-	public static function addTitle(&$object, $label, $level, $rang=-1)
+	public static function addTitle(&$object, $label, $level, $rang=-1, $desc = '')
 	/**
 	 * @param CommonObject $object
 	 * @param string       $label
 	 * @param int          $level
 	 * @param int          $rang
+	 * @param string       $desc
 	 * @return int
 	 */
 	{
-		return self::addSubTotalLine($object, $label, $level, $rang);
+		return self::addSubTotalLine($object, $label, $level, $rang, $desc);
 	}
 
 	public static function addTotal(&$object, $label, $level, $rang=-1)